+++ Rezolver.sln
@@ -65,20 +65,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.Microsoft.Extensio
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.Microsoft.AspNetCore.Hosting", "src\Rezolver.Microsoft.AspNetCore.Hosting\Rezolver.Microsoft.AspNetCore.Hosting.csproj", "{D88A4858-C6EF-4B85-80B2-293C9C031272}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "IOCScript", "IOCScript", "{775A6D70-BDA1-465F-AD71-C2B79F0DA31E}"
-	ProjectSection(SolutionItems) = preProject
-		src\IOCScript\readme.md = src\IOCScript\readme.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rezolver.IOCScript.Codegen", "src\IOCScript\Rezolver.IOCScript.Codegen\Rezolver.IOCScript.Codegen.csproj", "{1218A304-E417-4E37-8272-6C60703556EA}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.IOCScript", "src\IOCScript\Rezolver.IOCScript\Rezolver.IOCScript.csproj", "{B1063100-5EC8-486E-B6D2-53B2509540B5}"
-	ProjectSection(ProjectDependencies) = postProject
-		{1218A304-E417-4E37-8272-6C60703556EA} = {1218A304-E417-4E37-8272-6C60703556EA}
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.IOCScript.Tests", "test\Rezolver.IOCScript.Tests\Rezolver.IOCScript.Tests.csproj", "{3EB2EE85-37BB-408F-8137-48AB235E62C9}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.Benchmark", "test\Rezolver.Benchmark\Rezolver.Benchmark.csproj", "{BFE162A7-CD0B-4603-AF33-87C76B0A75E3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rezolver.Docs", "doc\Rezolver.Docs\Rezolver.Docs.csproj", "{E7055E56-9E6A-449E-87A5-35A31F652548}"
@@ -169,27 +155,6 @@ Global
 		{D88A4858-C6EF-4B85-80B2-293C9C031272}.PerformanceTesting|Any CPU.Build.0 = PerformanceTesting|Any CPU
 		{D88A4858-C6EF-4B85-80B2-293C9C031272}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D88A4858-C6EF-4B85-80B2-293C9C031272}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.Documentation|Any CPU.ActiveCfg = Release|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.PerformanceTesting|Any CPU.ActiveCfg = PerformanceTesting|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.PerformanceTesting|Any CPU.Build.0 = PerformanceTesting|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1218A304-E417-4E37-8272-6C60703556EA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.Documentation|Any CPU.ActiveCfg = Release|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.PerformanceTesting|Any CPU.ActiveCfg = PerformanceTesting|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.PerformanceTesting|Any CPU.Build.0 = PerformanceTesting|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1063100-5EC8-486E-B6D2-53B2509540B5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.Documentation|Any CPU.ActiveCfg = Release|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.PerformanceTesting|Any CPU.ActiveCfg = PerformanceTesting|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.PerformanceTesting|Any CPU.Build.0 = PerformanceTesting|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BFE162A7-CD0B-4603-AF33-87C76B0A75E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BFE162A7-CD0B-4603-AF33-87C76B0A75E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BFE162A7-CD0B-4603-AF33-87C76B0A75E3}.Documentation|Any CPU.ActiveCfg = Release|Any CPU
@@ -236,10 +201,6 @@ Global
 		{4534E330-B791-48F4-8A82-0C97D5F64322} = {C6B597B0-5AF4-43FD-AD4F-4A28CE0C0CDD}
 		{52399C04-A89B-44F5-B5B4-F123E9FDE80B} = {27A5E07B-3BEC-4F3A-8215-A9FF592F5F14}
 		{D88A4858-C6EF-4B85-80B2-293C9C031272} = {27A5E07B-3BEC-4F3A-8215-A9FF592F5F14}
-		{775A6D70-BDA1-465F-AD71-C2B79F0DA31E} = {27A5E07B-3BEC-4F3A-8215-A9FF592F5F14}
-		{1218A304-E417-4E37-8272-6C60703556EA} = {775A6D70-BDA1-465F-AD71-C2B79F0DA31E}
-		{B1063100-5EC8-486E-B6D2-53B2509540B5} = {775A6D70-BDA1-465F-AD71-C2B79F0DA31E}
-		{3EB2EE85-37BB-408F-8137-48AB235E62C9} = {C6B597B0-5AF4-43FD-AD4F-4A28CE0C0CDD}
 		{BFE162A7-CD0B-4603-AF33-87C76B0A75E3} = {C6B597B0-5AF4-43FD-AD4F-4A28CE0C0CDD}
 		{E7055E56-9E6A-449E-87A5-35A31F652548} = {73996028-9FE2-4094-BB1E-B2FDDB1CFB8F}
 		{8617C0D5-4EF8-46DC-9F9D-A516EA0B86FE} = {27A5E07B-3BEC-4F3A-8215-A9FF592F5F14}
