--- src/IOCScript/Rezolver.IOCScript/AssemblyInfo.cs
@@ -1,1 +1,0 @@
-﻿[assembly: System.CLSCompliant(true)]
