--- src/IOCScript/Rezolver.IOCScript/Class1.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Reflection;
-
-namespace Rezolver.IOCScript
-{
-    public class Class1
-    {
-        
-    }
-}
