--- src/IOCScript/Rezolver.IOCScript/Rezolver.IOCScript.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard1.1;net451;net46</TargetFrameworks>
-    <Configurations>Debug;Release;PerformanceTesting</Configurations>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\Rezolver.IOCScript.Codegen\obj\$(Configuration)\iocscript*.cs">
-      <Link>Parser\%(Filename)%(Extension)</Link>
-    </Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Antlr4.Runtime" Version="4.6.4" />
-  </ItemGroup>
-</Project>
