+++ src/Rezolver.Microsoft.AspNetCore.Hosting/Rezolver.Microsoft.AspNetCore.Hosting.csproj
@@ -3,16 +3,15 @@
   <PropertyGroup>
     <AssemblyName>Rezolver.Microsoft.AspNetCore.Hosting</AssemblyName>
     <AssemblyTitle>Rezolver Asp.Net Core application extensions</AssemblyTitle>
-    <Description>Extensions for IWebHostBuilder to simplify the use of Rezolver containers in Asp.Net Core 2 applications</Description>
+    <Description>Extensions for IWebHostBuilder to simplify the use of Rezolver containers in Asp.Net Core 2 applications.  This is the last version of this package.</Description>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-
     <PackageId>Rezolver.Microsoft.AspNetCore.Hosting</PackageId>
     <PackageTags>ioc;di;dependency injection;rezolver;dotnetcore;aspnetcore</PackageTags>
     <Title>Rezolver Asp.Net Core Extensions</Title>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <PackageOutputPath>../../.nupkg</PackageOutputPath>
-    <VersionPrefix>2.2.1</VersionPrefix>
+    <VersionPrefix>2.3</VersionPrefix>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <AllowedOutputExtensionsInPackageBuildOutputFolder>$(AllowedOutputExtensionsInPackageBuildOutputFolder);.pdb</AllowedOutputExtensionsInPackageBuildOutputFolder>
@@ -29,14 +28,15 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Routing" Version="2.2.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Routing" Version="2.2.2" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="2.2.0" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta-62925-02" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta2-19367-01" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup>
     <Folder Include="Properties\" />
+    <None Include="icon64x64.png" Pack="true" PackagePath="\" />
   </ItemGroup>
 
 </Project>
