+++ src/Rezolver.Microsoft.Extensions.Hosting/Rezolver.Microsoft.Extensions.Hosting.csproj
@@ -11,7 +11,7 @@
     <Title>Rezolver Generic Host Support</Title>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <PackageOutputPath>../../.nupkg</PackageOutputPath>
-    <VersionPrefix>2.2.1</VersionPrefix>
+    <VersionPrefix>3.0</VersionPrefix>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <AllowedOutputExtensionsInPackageBuildOutputFolder>$(AllowedOutputExtensionsInPackageBuildOutputFolder);.pdb</AllowedOutputExtensionsInPackageBuildOutputFolder>
@@ -25,7 +25,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -33,4 +33,7 @@
     <ProjectReference Include="..\Rezolver\Rezolver.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Include="icon64x64.png" Pack="true" PackagePath="\"/>
+  </ItemGroup>
 </Project>
