+++ src/Rezolver/Rezolver.csproj
@@ -4,8 +4,9 @@
     <AssemblyName>Rezolver</AssemblyName>
     <AssemblyTitle>Rezolver Core</AssemblyTitle>
     <Description>Rezolver is a fast IOC framework that lets you work the way you want to.  It's easy to use and very extensible.</Description>
-    <TargetFrameworks>netstandard2.0;netcoreapp2.2;net472;net48</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;netcoreapp2.2;net472;net48;netstandard2.1</TargetFrameworks>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <VersionPrefix>2.2.0</VersionPrefix>
     <PackageId>Rezolver</PackageId>
     <PackageTags>ioc;di;dependency injection;rezolver;dotnetcore</PackageTags>
     <Title>Rezolver IOC</Title>
@@ -33,14 +34,15 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta-62925-02" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta2-19367-01" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(EnableILEmit)' == 'True'">
-    <PackageReference Include="System.Reflection.Emit" Version="4.3.0" />
+    <PackageReference Include="System.Reflection.Emit" Version="4.6.0" />
   </ItemGroup>
 
   <ItemGroup>
+    <None Include="icon64x64.png" Pack="true" PackagePath="\" />
     <None Update="RootTargetContainerExtensions.AutoFactory.tt">
       <Generator>TextTemplatingFileGenerator</Generator>
       <LastGenOutput>RootTargetContainerExtensions.AutoFactory.cs</LastGenOutput>
