+++ test/Rezolver.Benchmark/Rezolver.Benchmark.csproj
@@ -8,8 +8,8 @@
     <Configurations>Debug;Release;PerformanceTesting</Configurations>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.11.4" />
-    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.11.4" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.11.5" />
+    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.11.5" />
   </ItemGroup>
 
   <ItemGroup>
