--- test/Rezolver.IOCScript.Tests/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Rezolver.IOCScript.Tests
-{
-    public class Class1
-    {
-    }
-}
