--- test/Rezolver.IOCScript.Tests/Rezolver.IOCScript.Tests.csproj
@@ -1,26 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AssemblyName>Rezolver.IOCScript.Tests</AssemblyName>
-    <PackageId>Rezolver.IOCScript.Tests</PackageId>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-    <Configurations>Debug;Release;PerformanceTesting</Configurations>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <DebugType>embedded</DebugType>
-    <DebugSymbols>True</DebugSymbols>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
-    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.0" />
-  </ItemGroup>
-
-</Project>
