--- test/Rezolver.Tests.Configuration.Json/xunit.runner.json
@@ -1,3 +1,0 @@
-﻿{
-  "methodDisplay": "method"
-}
